seperated all functions into correct controllers and models files
