+++ __tests__/app.test.js
@@ -143,3 +143,23 @@ describe("PATCH /api/artices/:articled_id", () => {
       });
   });
 });
+
+describe("GET /api/users", () => {
+  test("200: responds with an array of user objects", () => {
+    return request(app)
+      .get("/api/users")
+      .expect(200)
+      .then(({ body }) => {
+        const { topics } = body;
+        expect(topics).toBeInstanceOf(Array);
+        expect(topics).toHaveLength(4);
+        topics.forEach((topic) => {
+          expect(topic).toMatchObject({
+            username: expect.any(String),
+            name: expect.any(String),
+            avatar_url: expect.any(String),
+          });
+        });
+      });
+  });
+});
