+++ controllers/articles.controllers.js
@@ -1,0 +1,27 @@
+const {
+  updateArticle,
+  selectArticleById,
+} = require("../models/articles.model");
+
+exports.getArticleById = (req, res, next) => {
+  const { article_id } = req.params;
+  selectArticleById(article_id)
+    .then((article) => {
+      res.status(200).send({ article });
+    })
+    .catch((err) => {
+      next(err);
+    });
+};
+
+exports.patchArticle = (req, res, next) => {
+  const { article_id } = req.params;
+  const input = req.body;
+  updateArticle(article_id, input)
+    .then((article) => {
+      res.status(201).send({ article });
+    })
+    .catch((err) => {
+      next(err);
+    });
+};
