+++ controllers/topics.controllers.js
@@ -1,0 +1,11 @@
+const { selectTopics } = require("../models/topics.models");
+
+exports.getTopics = (req, res, next) => {
+  selectTopics()
+    .then((topics) => {
+      res.status(200).send({ topics });
+    })
+    .catch((err) => {
+      next(err);
+    });
+};
